+++ src/avt/DBAtts/SIL/avtSILEnumeratedNamespace.h
@@ -76,7 +76,7 @@ class DBATTS_API avtSILEnumeratedNamespace : public avtSILNamespace
     virtual            ~avtSILEnumeratedNamespace() {;};
 
     virtual int                      GetNumberOfElements(void) const
-                                                 { return elements.size(); };
+                                                 { return static_cast<int>(elements.size()); };
     virtual int                      GetElement(int idx) const
                                                  { return elements[idx]; };
     virtual NamespaceAttributes     *GetAttributes(void) const;
