+++ src/avt/Pipeline/Data/avtCompositeRF.C
@@ -180,8 +180,8 @@ avtCompositeRF::GetRayValue(const avtRay *ray,
     const float  *sample      = ray->sample[colorVariableIndex];
     const float  *sample2     = ray->sample[opacityVariableIndex];
     const float  *weight      = NULL;
-    float         min_weight  = 0.;
-    float         min_weight_denom = 0.;
+    double        min_weight  = 0.;
+    double        min_weight_denom = 0.;
     if (weightVariableIndex >= 0)
     {
         weight = ray->sample[weightVariableIndex];
