+++ src/avt/Pipeline/Data/avtDataAttributes.C
@@ -249,7 +249,7 @@ avtDataAttributes::avtDataAttributes() : plotInfoAtts()
     {
         for (int j=0; j<3; j++)
         {
-            unitCellVectors[i*3+j] = (i==j) ? 1.0 : 0.0;
+            unitCellVectors[i*3+j] = (i==j) ? 1.f : 0.f;
         }
     }
     unitCellOrigin[0] = unitCellOrigin[1] = unitCellOrigin[2] = 0.0;
@@ -2788,7 +2788,7 @@ avtDataAttributes::Write(avtDataObjectString &str,
     int   i, j;
 
     int varSize = 7;
-    int numVals = 35 + varSize*variables.size();
+    int numVals = 35 + static_cast<int>(varSize*variables.size());
     int *vals = new int[numVals];
     i = 0;
     vals[i++] = topologicalDimension;
@@ -2825,7 +2825,7 @@ avtDataAttributes::Write(avtDataObjectString &str,
     vals[i++] = (rectilinearGridHasTransform ? 1 : 0);
     vals[i++] = (constructMultipleCurves ? 1 : 0);
     vals[i++] = activeVariable;
-    vals[i++] = variables.size();
+    vals[i++] = static_cast<int>(variables.size());
     int basei = i;
     for (i = 0 ; i < variables.size() ; i++)
     {
@@ -2833,8 +2833,8 @@ avtDataAttributes::Write(avtDataObjectString &str,
         vals[basei+varSize*i+1] = variables[i]->centering;
         vals[basei+varSize*i+2] = (variables[i]->treatAsASCII ? 1 : 0);
         vals[basei+varSize*i+3] = variables[i]->vartype;
-        vals[basei+varSize*i+4] = variables[i]->subnames.size();
-        vals[basei+varSize*i+5] = variables[i]->binRange.size();
+        vals[basei+varSize*i+4] = static_cast<int>(variables[i]->subnames.size());
+        vals[basei+varSize*i+5] = static_cast<int>(variables[i]->binRange.size());
         vals[basei+varSize*i+6] = variables[i]->useForAxis;
     }
     wrtr->WriteInt(str, vals, numVals);
@@ -2849,13 +2849,13 @@ avtDataAttributes::Write(avtDataObjectString &str,
     for (i = 0 ; i < variables.size() ; i++)
     {
         // Write the variable name
-        wrtr->WriteInt(str, variables[i]->varname.size());
+        wrtr->WriteInt(str, static_cast<int>(variables[i]->varname.size()));
         str.Append((char *) variables[i]->varname.c_str(),
-                   variables[i]->varname.size(),
+                   static_cast<int>(variables[i]->varname.size()),
                    avtDataObjectString::DATA_OBJECT_STRING_SHOULD_MAKE_COPY);
 
         // Write the units name.
-        int unitlen = variables[i]->varunits.size();
+        int unitlen = static_cast<int>(variables[i]->varunits.size());
         wrtr->WriteInt(str, unitlen);
         if(unitlen > 0)
         {
@@ -2867,9 +2867,9 @@ avtDataAttributes::Write(avtDataObjectString &str,
         // communicated in mass "int" writing phase.
         for (j = 0 ; j < variables[i]->subnames.size() ; j++)
         {
-            wrtr->WriteInt(str, variables[i]->subnames[j].size());
+            wrtr->WriteInt(str, static_cast<int>(variables[i]->subnames[j].size()));
             str.Append((char *) variables[i]->subnames[j].c_str(),
-                     variables[i]->subnames[j].size(),
+                     static_cast<int>(variables[i]->subnames[j].size()),
                      avtDataObjectString::DATA_OBJECT_STRING_SHOULD_MAKE_COPY);
         }
         // Write the binRanges (if any).  Number of binRanges already
@@ -2877,7 +2877,7 @@ avtDataAttributes::Write(avtDataObjectString &str,
         if (variables[i]->binRange.size() > 0)
         {
             wrtr->WriteDouble(str, &(variables[i]->binRange[0]), 
-                              variables[i]->binRange.size());
+                              static_cast<int>(variables[i]->binRange.size()));
         }
         variables[i]->originalData->Write(str, wrtr);
         variables[i]->thisProcsOriginalData->Write(str, wrtr);
