+++ src/common/misc/TimingsManager.h
@@ -192,7 +192,7 @@ class MISC_API SystemTimingsManager : public TimingsManager
     void                       ClearValues()
                                    { values.clear(); };
     int                        GetNValues() const
-                                   { return values.size(); };
+                                   { return static_cast<int>(values.size()); };
 
   protected:
     std::vector<struct TIMEINFO> values;
@@ -208,7 +208,7 @@ class MISC_API MPITimingsManager : public TimingsManager
     void                       ClearValues()
                                    { values.clear(); };
     int                        GetNValues() const
-                                   { return values.size(); };
+                                   { return static_cast<int>(values.size()); };
 
   protected:
     std::vector<double>        values;
